transform action email to core net
