+++ Ginger/GingerCoreNETUnitTest/LinuxTransformationTests/ActeMailTests.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace GingerCoreNETUnitTest.LinuxTransformationTests
+{
+    class ActeMailTests
+    {
+    }
+}
